Split up sidecar into files
